encoders ported, 347 tests passed, 7 failed
